cleanup old donotanswer txt files, move tests into own module
